backend/requirements.txt
# ── Core Framework ────────────────────────────────────────────────────────────
# Pinned for Python 3.13 Windows compatibility (no Rust compilation needed)
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.10.3
pydantic-settings==2.6.1

# ── Database ──────────────────────────────────────────────────────────────────
motor==3.6.0
pymongo==4.10.1

# ── Google APIs ───────────────────────────────────────────────────────────────
google-generativeai==0.8.3
google-auth==2.37.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.154.0

# ── Cryptography & Security ───────────────────────────────────────────────────
cryptography==44.0.0

# ── HTTP Client ───────────────────────────────────────────────────────────────
httpx==0.28.1

# ── File Handling ─────────────────────────────────────────────────────────────
python-multipart==0.0.17
pdfplumber==0.11.4
python-docx==1.1.2

# ── Utilities ─────────────────────────────────────────────────────────────────
python-dotenv==1.0.1